feat: remove auxiliaryBarTab
