+++ src/models/layout.ts
@@ -6,7 +6,7 @@ export enum LayoutEvents {
 
 export class LayoutModel {
     constructor(
-        public splitPanePos: PosType[] = [300, 'auto', 300],
+        public splitPanePos: PosType[] = [300, 'auto', 25],
         public horizontalSplitPanePos: PosType[] = ['auto', 230],
         public groupSplitPos: number[] = [],
         public activityBar: WithHidden<void> = { hidden: false },
